requirements.txt
asgiref==3.7.2
Babel==2.12.1
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docx==0.2.4
docxcompose==1.4.0
docxtpl==0.16.7
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
Pillow==10.0.0
psycopg2==2.9.7
PyJWT==2.8.0
python-docx==0.8.11
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
